+++ lib/sqlite_db.dart
@@ -1,0 +1,3 @@
+library dbutils;
+
+export 'package:dbutils/src/sqlite/sqlite_db.dart';