+++ lib/sqllitedb.dart
@@ -1,3 +1,0 @@
-library dbutils;
-
-export 'src/sqlite/DBInterface.dart';